+++ packages/sit-onyx/src/components/OnyxCheckboxGroup/OnyxCheckboxGroup.stories.ts
@@ -30,6 +30,7 @@ export const Default = {
       { label: "Initially checked", id: "id-2" },
       { label: "Required", id: "id-3", required: true },
       { label: "Disabled", id: "id-4", disabled: true },
+      { label: "Loading", id: "id-5", loading: true },
     ],
   },
 } satisfies Story;