+++ packages/sit-onyx/src/components/OnyxRadioButtonGroup/OnyxRadioButtonGroup.stories.ts
@@ -23,6 +23,7 @@ const EXAMPLE_OPTIONS: SelectionOption<string>[] = [
   { label: "Default", value: "1", id: "1" },
   { label: "Initially checked", value: "2", id: "2" },
   { label: "Disabled", value: "3", id: "3", disabled: true },
+  { label: "Loading", value: "4", id: "4", loading: true },
 ];
 
 /**
@@ -69,10 +70,9 @@ export const WithTruncation = {
     ...Default.args,
     options: [
       { label: "Very long label that will be truncated", id: "id-1" },
-      { label: "Very long required label that will be truncated", id: "id-2" },
       {
         label: "Very long label that will be truncated with multiline",
-        id: "id-3",
+        id: "id-2",
         truncation: "multiline",
       },
     ],