+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,26 @@
 # alpha-test-app
 
+## 0.0.1-alpha.46
+
+### Patch Changes
+
+- Updated dependencies [668d5fe]
+  - sit-onyx@1.0.0-alpha.49
+
+## 0.0.1-alpha.45
+
+### Patch Changes
+
+- Updated dependencies [2eb0867]
+  - sit-onyx@1.0.0-alpha.48
+
+## 0.0.1-alpha.44
+
+### Patch Changes
+
+- Updated dependencies [9ad6208]
+  - sit-onyx@1.0.0-alpha.47
+
 ## 0.0.1-alpha.43
 
 ### Patch Changes