+++ packages/chartjs-plugin/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sit-onyx/chartjs-plugin
 
+## 1.0.0-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [668d5fe]
+  - sit-onyx@1.0.0-alpha.49
+
+## 1.0.0-alpha.2
+
+### Patch Changes
+
+- Updated dependencies [2eb0867]
+  - sit-onyx@1.0.0-alpha.48
+
+## 0.0.1-alpha.1
+
+### Patch Changes
+
+- Updated dependencies [9ad6208]
+  - sit-onyx@1.0.0-alpha.47
+
 ## 0.0.1-alpha.0
 
 ### Patch Changes