+++ packages/storybook-utils/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sit-onyx/storybook-utils
 
+## 1.0.0-alpha.51
+
+### Patch Changes
+
+- Updated dependencies [668d5fe]
+  - sit-onyx@1.0.0-alpha.49
+
+## 1.0.0-alpha.50
+
+### Patch Changes
+
+- Updated dependencies [2eb0867]
+  - sit-onyx@1.0.0-alpha.48
+
+## 1.0.0-alpha.49
+
+### Patch Changes
+
+- Updated dependencies [9ad6208]
+  - sit-onyx@1.0.0-alpha.47
+
 ## 1.0.0-alpha.48
 
 ### Patch Changes