+++ packages/vitepress-theme/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sit-onyx/vitepress-theme
 
+## 1.0.0-alpha.48
+
+### Patch Changes
+
+- Updated dependencies [668d5fe]
+  - sit-onyx@1.0.0-alpha.49
+
+## 1.0.0-alpha.47
+
+### Patch Changes
+
+- Updated dependencies [2eb0867]
+  - sit-onyx@1.0.0-alpha.48
+
+## 1.0.0-alpha.46
+
+### Patch Changes
+
+- Updated dependencies [9ad6208]
+  - sit-onyx@1.0.0-alpha.47
+
 ## 1.0.0-alpha.45
 
 ### Patch Changes